+++ src/Reflection/AbstractEntity.php
@@ -28,22 +28,37 @@ use phpDocumentor\Reflection\DocBlockFactory;
 abstract class AbstractEntity extends ParentAbstractEntity
 {
     /**
+     * @var \phpDocumentor\Reflection\DocBlock|null
+     */
+    private $docBlock;
+
+    /**
+     * @var \Cake\Collection\Collection
+     */
+    private $tags;
+
+    /**
      * Internal method to get the `DocBlock` instance
-     * @param \Roave\BetterReflection\Reflection\Reflection|\Roave\BetterReflection\Reflection\ReflectionFunctionAbstract|null $reflectionObject A `Reflection` object
      * @return \phpDocumentor\Reflection\DocBlock|null
      */
-    protected function getDocBlockInstance($reflectionObject = null): ?DocBlock
+    protected function getDocBlockInstance(): ?DocBlock
     {
+        if ($this->docBlock) {
+            return $this->docBlock;
+        }
+
         try {
-            return DocBlockFactory::createInstance()->create($reflectionObject ?: $this->reflectionObject);
+            $this->docBlock = DocBlockFactory::createInstance()->create($this->reflectionObject);
         } catch (InvalidArgumentException $e) {
             //The exception will still be throwned in case of a malformed tag
             if (string_contains($e->getMessage(), 'does not seem to be wellformed, please check it for errors')) {
                 throw $e;
             }
 
-            return null;
+            $this->docBlock = null;
         }
+
+        return $this->docBlock;
     }
 
     /**
@@ -52,9 +67,7 @@ abstract class AbstractEntity extends ParentAbstractEntity
      */
     public function getDocBlockSummaryAsString(): string
     {
-        $DocBlockInstance = $this->getDocBlockInstance();
-
-        return $DocBlockInstance ? $DocBlockInstance->getSummary() : '';
+        return $this->getDocBlockInstance() ? $this->getDocBlockInstance()->getSummary() : '';
     }
 
     /**
@@ -63,9 +76,7 @@ abstract class AbstractEntity extends ParentAbstractEntity
      */
     public function getDocBlockDescriptionAsString(): string
     {
-        $DocBlockInstance = $this->getDocBlockInstance();
-
-        return $DocBlockInstance ? $DocBlockInstance->getDescription()->render() : '';
+        return $this->getDocBlockInstance() ? (string)$this->getDocBlockInstance()->getDescription() : '';
     }
 
     /**
@@ -85,26 +96,20 @@ abstract class AbstractEntity extends ParentAbstractEntity
     }
 
     /**
-     * Internal method to parse tags
-     * @param array $tags An array of `Tag` instances
-     * @return \Cake\Collection\Collection A `Collection` of `TagEntity` instances
-     */
-    protected function parseTags(array $tags): Collection
-    {
-        return collection($tags)->map(function (Tag $tag) {
-            return new TagEntity($tag);
-        }, $tags)->sortBy('name', SORT_ASC, SORT_STRING);
-    }
-
-    /**
      * Gets all tags
      * @return \Cake\Collection\Collection
      */
     public function getTags(): Collection
     {
-        $DocBlockInstance = $this->getDocBlockInstance();
+        if (!$this->tags instanceof Collection) {
+            $tags = $this->getDocBlockInstance() ? $this->getDocBlockInstance()->getTags() : [];
 
-        return $this->parseTags($DocBlockInstance ? $DocBlockInstance->getTags() : []);
+            $this->tags = collection($tags)->map(function (Tag $tag) {
+                return new TagEntity($tag);
+            })->sortBy('name', SORT_ASC, SORT_STRING);
+        }
+
+        return $this->tags;
     }
 
     /**
@@ -123,9 +128,7 @@ abstract class AbstractEntity extends ParentAbstractEntity
      */
     public function getTagsByName(string $name): Collection
     {
-        $DocBlockInstance = $this->getDocBlockInstance();
-
-        return $this->parseTags($DocBlockInstance ? $DocBlockInstance->getTagsByName($name) : []);
+        return $this->getTags()->match(compact('name'));
     }
 
     /**
@@ -135,8 +138,6 @@ abstract class AbstractEntity extends ParentAbstractEntity
      */
     public function hasTag(string $name): bool
     {
-        $DocBlockInstance = $this->getDocBlockInstance();
-
-        return $DocBlockInstance ? $DocBlockInstance->hasTag($name) : false;
+        return $this->getDocBlockInstance() ? $this->getDocBlockInstance()->hasTag($name) : false;
     }
 }