+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 1.x branch
 ## 1.0 branch
 ### 1.0.3-beta4
+* improved `getDocBlockInstance()` and `getTags()` method provided by the
+    `AbstractEntity`; The values are now cached;
 * added `functions.rendering` and `class.rendering` events, dispatched by
     `PhpDocMaker`. This allows easier identification of errors.